ITransformHierarchyObject interface removed.
