+++ ShapeEngine/Core/Shapes/Shape.cs
@@ -1,12 +1,9 @@
 using System.Numerics;
 using ShapeEngine.Core.Collision;
+using ShapeEngine.Core.Interfaces;
 using ShapeEngine.Core.Structs;
 
 namespace ShapeEngine.Core.Shapes;
-
-//a way to change absolute position/size of a child
-//calculates offset based on cur transform and new absolute transform!!!
-
 
 public abstract class Shape : IShape
 {
@@ -17,7 +14,7 @@ public abstract class Shape : IShape
     public Transform2D Offset { get; set; }
     public Transform2D CurTransform { get; private set; }
     public Transform2D PrevTransform { get; private set; }
-    
+
     
     protected Shape()
     {
@@ -102,3 +99,48 @@ protected virtual void Update(float dt) { }
 protected virtual void Draw() { }
 protected virtual void OnTransformSetupFinished() { }
 */
+
+
+/*
+public interface ITransformHierarchyObject
+{
+    public Transform2D CurTransform { get; set; }
+    
+    /// <summary>
+    /// The offset used for children only.
+    /// </summary>
+    public Transform2D Offset { get; set; }
+    public List<ITransformHierarchyObject> GetChildren();
+    
+    public bool IsMoving() => true;
+    public bool IsRotating() => true;
+    public bool IsScaling() => true;
+
+
+    /// <summary>
+    /// Called on the parent to update all children.
+    /// </summary>
+    /// <param name="transform">The transform for this instance.</param>
+    public sealed void UpdateHierarchy(Transform2D transform)
+    {
+        UpdateChildren(transform);
+    }
+
+    public void OnTransformUpdated(Transform2D prevTransform, Transform2D newTransform);
+    private void UpdateChildren(Transform2D parentTransform)
+    {
+        var children = GetChildren();
+        var prevTransform = CurTransform;
+        
+        CurTransform = Transform2D.UpdateTransform(parentTransform, Offset, IsMoving(), IsRotating(), IsScaling());
+        
+        OnTransformUpdated(prevTransform, CurTransform);
+        
+        foreach (var child in children)
+        {
+            child.UpdateChildren(CurTransform);
+        }
+    }
+}
+*/
+
